+++ src/components/Modal.js
@@ -1,0 +1,1 @@
+import Modal from 'react-bootstrap/Button';
